requirements.txt
streamlit>=1.28.0
pypdf2>=3.0.1
python-pptx>=0.6.21
pytesseract>=0.3.10
Pillow>=10.0.0
openai>=1.0.0
python-dotenv>=1.0.0
